Refactor Resizer
